summarization_backend/summarizer_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==12.0.0
docx2txt==0.8
PyPDF2==3.0.1
sumy==0.11.0
nltk==3.8.1
transformers==4.30.2
torch==2.0.1
